attempt WindSpeed - incomplete

+++ app/controllers/StationCtrl.java
@@ -22,7 +22,7 @@ public class StationCtrl extends Controller
         render("station.html", station);
     }
 
-    public static void addReading(Long id, int code, double temperature, double windSpeed, int pressure) {
+    public static void addReading(Long id, int code, double temperature, int windSpeed, int pressure) {
         Reading reading = new Reading(code, temperature, windSpeed, pressure);
         Station station = Station.findById(id);
         station.readings.add(reading);
